xephyr_isolate_from_root
Having just one domain for nested xephyr instances makes the xnested sandboxing
method of limited utility when used for more than a few sensitive apps.

So, here is a smaller template that only constrains apps from the root xserver.
Making specific xephyr domains is left to other means.

I intend to make a template for such domains in the future, but not for now.